Fixed: Dynamic memory management issue with Assignee
The memory management issue was happening due to implicitly deleted instance of TransferGroupInfo. To prevent that, created dynamically loaded TransferGroupInfo ('new') and kept it using a copy contructor of it. Now the issue is AssigneeInfo do not seem to contain information that it should contain. It could be because I did not override the copy contructor, something TransferGroupInfo currently has.