Remove pointers and move to copy contructors
The good thing about copy contructors is that they are always memery-leak-safe when it comes to traveling objects. As long as the instance can be kept as a copy and unless the callback is not required to the same object, it is okay. They are removed when they are out of scope which is a good thing. And I always forget that nullptr is for pointers (could have seen thing coming since it ends with ptr (pointer)). Other than that I am beginning to understand why references once initialized cannot be changed or be initialized as null objects using nullptr as such.